+++ changelog.md
@@ -3,6 +3,7 @@
 #### v3.3.0
 
 * `[system/process]` Added new package for getting information about active system processes
+* `[terminal]` Fixed bug with title formating in `ReadAnswer` function
 
 #### v3.2.3